+++ Classes/Geocoding/Google.php
@@ -126,11 +126,14 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
         }
 
         $address = $geoObject->getGeoAddress();
+        $configuration = Tx_Oelib_ConfigurationRegistry::get('plugin.tx_oelib');
+        $apiKey = $configuration->getAsString('googleGeocodingApiKey');
+        $url = self::BASE_URL . '&key=' . \urlencode($apiKey) . '&address=' . \urlencode($address);
         $delayInMicroseconds = self::INITIAL_DELAY_IN_MICROSECONDS;
 
         while (true) {
             \usleep($delayInMicroseconds);
-            $response = $this->sendRequest($address);
+            $response = $this->sendRequest($url);
             if ($response === false) {
                 $status = 'General network problem.';
             } else {
@@ -166,14 +169,12 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
     /**
      * Sends a geocoding request to the Google Maps server.
      *
-     * @param string $address the address to look up, must not be empty
+     * @param string $url
      *
      * @return string|bool string with the JSON result from the Google Maps server, or false if an error has occurred
      */
-    protected function sendRequest($address)
+    protected function sendRequest($url)
     {
-        $baseUrlWithAddress = self::BASE_URL . '&address=';
-
-        return GeneralUtility::getUrl($baseUrlWithAddress . \urlencode($address));
+        return GeneralUtility::getUrl($url);
     }
 }