+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a configuration for the geocoding API key (#91, #98)
 - Add the possibility to set a reply-to address for a mail object (#86)
 
 ### Changed